refactor: modularize app components
